Changed directory for homework
